Split mission tests
